UIE-200 Ajax cleanup pt4  - broke out CromIAM sub area out of ajax.ts and into sub-module.  - only minimal conversion to Typescript.
